Edit: MET 1118 Edit modal compose report
