+++ src/components/StakingLifeCycle/DelegatorLifecycle/ReportComposerModal/index.tsx
@@ -1,29 +1,23 @@
 import { Box } from "@mui/material";
-import FilledInfoModal from "./FilledInfoModal";
-import StepTransferModal from "./StepTransferModal";
-import StepEventsModal from "./StepEventsModal";
-import StepReviewModal from "./StepReviewModal";
 import { useCallback, useEffect, useState } from "react";
+import FilledInfoModal from "./FilledInfoModal";
+import StepReviewModal from "./StepReviewModal";
 
 export interface IPropsModal {
   open: boolean;
   handleCloseModal: () => void;
   saveParams?: (params: any) => void;
   gotoStep?: (step: number) => void;
-  defaultParams?: any;
+  params?: any;
 }
 
 export enum STEPS {
   step1 = 1,
   step2 = 2,
-  step3 = 3,
-  step4 = 4
 }
 
 const ReportComposerModal = ({ open, handleCloseModal }: IPropsModal) => {
-  const [paramsStep1, setParamsStep1] = useState({});
-  const [paramsStep2, setParamsStep2] = useState({});
-  const [paramsStep3, setParamsStep3] = useState({});
+  const [params, setParams] = useState({});
   const [currentStep, setCurrentStep] = useState<STEPS>();
 
   const gotoStep = useCallback((step?: STEPS) => {
@@ -37,37 +31,21 @@ const ReportComposerModal = ({ open, handleCloseModal }: IPropsModal) => {
       setCurrentStep(undefined);
     }
   }, [open]);
-
-  const defaultParams = [paramsStep1, paramsStep2, paramsStep3];
-
   return (
     <Box data-testid="report-compose-modal">
       <FilledInfoModal
         open={currentStep === STEPS.step1}
         handleCloseModal={handleCloseModal}
-        saveParams={setParamsStep1}
+        saveParams={setParams}
         gotoStep={gotoStep}
       />
-      <StepTransferModal
-        open={currentStep === STEPS.step2}
-        handleCloseModal={handleCloseModal}
-        saveParams={setParamsStep2}
-        gotoStep={gotoStep}
-        defaultParams={defaultParams}
-      />
-      <StepEventsModal
-        open={currentStep === STEPS.step3}
-        handleCloseModal={handleCloseModal}
-        saveParams={setParamsStep3}
-        gotoStep={gotoStep}
-        defaultParams={defaultParams}
-      />
-      <StepReviewModal
-        open={currentStep === STEPS.step4}
-        handleCloseModal={handleCloseModal}
-        defaultParams={defaultParams}
-        gotoStep={gotoStep}
-      />
+      {currentStep === STEPS.step2 && (
+        <StepReviewModal
+          open
+          handleCloseModal={handleCloseModal}
+          params={params}
+          gotoStep={gotoStep}
+        />)}
     </Box>
   );
 };
